requirements.txt
pynput
pystray
Pillow
pyperclip
